+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Forecast.hs
@@ -2,25 +2,12 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE RoleAnnotations #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Dijkstra.Forecast (
   DijkstraForecast (..),
-  mkDijkstraForecast,
-  dfPoolDistrL,
-  dfMaxBlockHeaderSizeL,
-  dfMaxBlockBodySizeL,
-  dfProtocolVersionL,
-  dfLeiosCommitteeL,
-  dfLeiosCommitteeSizeL,
-  dfLeiosQuorumStakeThresholdL,
-  dfLeiosAnnouncementPeriodLengthL,
-  dfLeiosVotePeriodLengthL,
-  dfLeiosDiffusionPeriodLengthL,
-  dfMaxEndorserBlockReferencesSizeL,
-  dfMaxEndorserBlockTxsSizeL,
-  dfMaxEndorserBlockExUnitsL,
-  dfMaxRefScriptSizePerEndorserBlockL,
+  DijkstraEraForecast (..),
 ) where
 
 import Cardano.Ledger.BaseTypes (Milliseconds32, ProtVer, UnitInterval)
@@ -43,7 +30,6 @@ import Cardano.Ledger.Dijkstra.PParams (
 import Cardano.Ledger.Dijkstra.State.CertState ()
 import Cardano.Ledger.Plutus.ExUnits (OrdExUnits)
 import Cardano.Ledger.Shelley.API.Forecast (
-  DijkstraEraForecast (..),
   EraForecast (..),
   Timeline (..),
  )
@@ -73,6 +59,20 @@ import NoThunks.Class (NoThunks (..))
 -- The Leios fields are fixed at an epoch boundary -- the committee is seated on
 -- the stake snapshot by the SNAP rule, the rest are protocol parameters -- so
 -- they are forecastable for the same reason the Praos fields are.
+
+-- | Additional forecast fields available only in Leios eras.
+class EraForecast era => DijkstraEraForecast era where
+  leiosCommitteeForecastL :: Lens' (Forecast t era) LeiosCommittee
+  leiosCommitteeSizeForecastL :: Lens' (Forecast t era) Word16
+  leiosQuorumStakeThresholdForecastL :: Lens' (Forecast t era) UnitInterval
+  leiosAnnouncementPeriodLengthForecastL :: Lens' (Forecast t era) Milliseconds32
+  leiosVotePeriodLengthForecastL :: Lens' (Forecast t era) Milliseconds32
+  leiosDiffusionPeriodLengthForecastL :: Lens' (Forecast t era) Milliseconds32
+  maxEndorserBlockReferencesSizeForecastL :: Lens' (Forecast t era) Word32
+  maxEndorserBlockTxsSizeForecastL :: Lens' (Forecast t era) Word32
+  maxEndorserBlockExUnitsForecastL :: Lens' (Forecast t era) OrdExUnits
+  maxRefScriptSizePerEndorserBlockForecastL :: Lens' (Forecast t era) Word32
+
 data DijkstraForecast (t :: Timeline) era = DijkstraForecast
   { dfPoolDistr :: !PoolDistr
   , dfMaxBlockHeaderSize :: !Word16